# [NOD-1387] Some small optimizations to block processing