these functions were pretty high in benchmarks of a test I'm writing to process thousands of blocks,
[CPU profile before](https://github.com/kaspanet/kaspad/files/5214203/cpu_before.pdf)
[Memory profile before](https://github.com/kaspanet/kaspad/files/5214205/mem_before.pdf)

[CPU profile after the clone changes](https://github.com/kaspanet/kaspad/files/5214210/cpu_clone_preallocate.pdf)
[Memory profile after the clone changes](https://github.com/kaspanet/kaspad/files/5214207/mem_clone_preallocate.pdf)

[CPU profile after this PR](https://github.com/kaspanet/kaspad/files/5214213/cpu_endian_better.pdf)
[Memory profile after this PR](https://github.com/kaspanet/kaspad/files/5214212/mem_endian_better.pdf)

I also wrote a new benchmark for `serializeUTXO` that benchmarks serializing a common UTXO.
before this PR:
```
goos: linux                                                                                                                                                                                                                                  
goarch: amd64                                           
pkg: github.com/kaspanet/kaspad/domain/blockdag
Benchmark_serializeUTXO-16       3623398               325 ns/op         
PASS                                                                                                                  
ok      github.com/kaspanet/kaspad/domain/blockdag      1.522s                                    
```
after:
```
goos: linux
goarch: amd64                                  
pkg: github.com/kaspanet/kaspad/domain/blockdag                                                                       
Benchmark_serializeUTXO-16      15396980                70.6 ns/op
PASS                                                       
ok      github.com/kaspanet/kaspad/domain/blockdag      1.177s
```